+++ pkg/maps/lbmap/maglev.go
@@ -104,7 +104,7 @@ func deleteMapIfMNotMatch(mapName string, tableSize uint32) (bool, error) {
 
 // updateMaglevTable creates a new inner Maglev map containing the given backend IDs
 // and sets it as the active lookup table for the given service ID.
-func updateMaglevTable(ipv6 bool, revNATID uint16, backendIDs []loadbalancer.BackendID) error {
+func updateMaglevTable(ipv6 bool, revNATID uint32, backendIDs []loadbalancer.BackendID) error {
 	outer := maglevOuter4Map
 	if ipv6 {
 		outer = maglevOuter6Map
@@ -132,7 +132,7 @@ func updateMaglevTable(ipv6 bool, revNATID uint16, backendIDs []loadbalancer.Bac
 }
 
 // deleteMaglevTable deletes the inner Maglev lookup table for the given service ID.
-func deleteMaglevTable(ipv6 bool, revNATID uint16) error {
+func deleteMaglevTable(ipv6 bool, revNATID uint32) error {
 	outerMap := maglevOuter4Map
 	if ipv6 {
 		outerMap = maglevOuter6Map
