+++ pkg/service/service.go
@@ -976,7 +976,7 @@ func (s *Service) UpdateBackendsState(backends []*lb.Backend) error {
 
 				if p, found = updateSvcs[id]; !found {
 					p = &datapathTypes.UpsertServiceParams{
-						ID:                        uint16(id),
+						ID:                        uint32(id),
 						IP:                        info.frontend.L3n4Addr.AddrCluster.AsNetIP(),
 						Port:                      info.frontend.L3n4Addr.L4Addr.Port,
 						PrevBackendsCount:         len(info.backends),
@@ -1381,7 +1381,7 @@ func (s *Service) deleteBackendsFromAffinityMatchMap(svcID lb.ID, backendIDs []l
 	}).Debug("Deleting backends from session affinity match")
 
 	for _, bID := range backendIDs {
-		if err := s.lbmap.DeleteAffinityMatch(uint16(svcID), bID); err != nil {
+		if err := s.lbmap.DeleteAffinityMatch(uint32(svcID), bID); err != nil {
 			log.WithFields(logrus.Fields{
 				logfields.BackendID: bID,
 				logfields.ServiceID: svcID,
@@ -1397,7 +1397,7 @@ func (s *Service) addBackendsToAffinityMatchMap(svcID lb.ID, backendIDs []lb.Bac
 	}).Debug("Adding backends to affinity match map")
 
 	for _, bID := range backendIDs {
-		if err := s.lbmap.AddAffinityMatch(uint16(svcID), bID); err != nil {
+		if err := s.lbmap.AddAffinityMatch(uint32(svcID), bID); err != nil {
 			log.WithFields(logrus.Fields{
 				logfields.BackendID: bID,
 				logfields.ServiceID: svcID,
@@ -1452,7 +1452,7 @@ func (s *Service) upsertServiceIntoLBMaps(svc *svcInfo, isExtLocal, isIntLocal b
 
 	// Update LB source range check cidrs
 	if checkLBSrcRange = svc.checkLBSourceRange() || len(prevLoadBalancerSourceRanges) != 0; checkLBSrcRange {
-		if err := s.lbmap.UpdateSourceRanges(uint16(svc.frontend.ID),
+		if err := s.lbmap.UpdateSourceRanges(uint32(svc.frontend.ID),
 			prevLoadBalancerSourceRanges, svc.loadBalancerSourceRanges,
 			v6FE); err != nil {
 
@@ -1508,7 +1508,7 @@ func (s *Service) upsertServiceIntoLBMaps(svc *svcInfo, isExtLocal, isIntLocal b
 	svc.svcNatPolicy = natPolicy
 
 	p := &datapathTypes.UpsertServiceParams{
-		ID:                        uint16(svc.frontend.ID),
+		ID:                        uint32(svc.frontend.ID),
 		IP:                        svc.frontend.L3n4Addr.AddrCluster.AsNetIP(),
 		Port:                      svc.frontend.L3n4Addr.L4Addr.Port,
 		PreferredBackends:         preferredBackends,
@@ -1752,7 +1752,7 @@ func (s *Service) restoreServicesLocked(svcBackendsById map[lb.BackendID]struct{
 
 				backends[b.String()] = b
 			}
-			if err := s.lbmap.UpsertMaglevLookupTable(uint16(newSVC.frontend.ID), backends,
+			if err := s.lbmap.UpsertMaglevLookupTable(uint32(newSVC.frontend.ID), backends,
 				ipv6); err != nil {
 				scopedLog.WithError(err).Warning("Unable to upsert into the Maglev BPF map.")
 				continue
@@ -1798,7 +1798,7 @@ func (s *Service) deleteServiceLocked(svc *svcInfo) error {
 
 	if option.Config.EnableSVCSourceRangeCheck &&
 		svc.svcType == lb.SVCTypeLoadBalancer {
-		if err := s.lbmap.UpdateSourceRanges(uint16(svc.frontend.ID),
+		if err := s.lbmap.UpdateSourceRanges(uint32(svc.frontend.ID),
 			svc.loadBalancerSourceRanges, nil, ipv6); err != nil {
 			return err
 		}
